+++ src/main/java/org/folio/qm/validation/SkippedValidationError.java
@@ -1,0 +1,6 @@
+package org.folio.qm.validation;
+
+import org.folio.rspec.validation.validator.marc.model.MarcRuleCode;
+
+public record SkippedValidationError(String tagName, MarcRuleCode ruleCode) {
+}